A couple of fixes to device check-out
The MobileApplicationProfile is not retrieved from BlueSphere for a new device until after the customer code has been entered.  At this point, calling ApplicationRepository.GetAsync() will throw an exception.  Therefore it is not possible to make this call in the NavigationService.SetMappings, which is executed from within the NavigationService's constructor, or from the StartupViewModel if there is no customer repository data.